+++ src/neo/Oracle/OracleRequest.cs
@@ -4,20 +4,12 @@ namespace Neo.Oracle
 {
     public abstract class OracleRequest
     {
-        /// <summary>
-        /// Only for hashing entropy
-        /// </summary>
-        public enum RequestType : byte
-        {
-            HTTPS = 0x01,
-        }
-
         private UInt160 _hash;
 
         /// <summary>
         /// Type
         /// </summary>
-        public RequestType Type { get; }
+        public OracleRequestType Type { get; }
 
         /// <summary>
         /// Hash
@@ -39,7 +31,7 @@ namespace Neo.Oracle
         /// Constructor
         /// </summary>
         /// <param name="type">Type</param>
-        protected OracleRequest(RequestType type) { Type = type; }
+        protected OracleRequest(OracleRequestType type) { Type = type; }
 
         /// <summary>
         /// This method serialize the parts of the class that should be taken into account for compute the Hash
