+++ src/neo/Oracle/OracleRequestType.cs
@@ -1,0 +1,7 @@
+namespace Neo.Oracle
+{
+    public enum OracleRequestType : byte
+    {
+        HTTPS = 0x01,
+    }
+}
